linked the external getDirections to pop up modals
