+++ ConcordiaMaps/components/MapMarkers.js
@@ -1,78 +1,41 @@
-import React, { useState } from "react";
-import { Marker, Callout } from "react-native-maps";
-import { Text, View, Image, StyleSheet } from "react-native";
-import PropTypes from "prop-types";
-import PopupModal from "./PopupModal";
-import { useNavigation, useRoute } from "@react-navigation/native";
+import React, { useContext } from "react";
+import { Marker } from "react-native-maps";
+import { Image } from "react-native";
 import styles from "../styles";
 import { Building } from "../constants/Building";
+import { ModalContext } from "../App";
 const customMarkerImage = require("../assets/PinLogo.png");
 
 const MapMarkers = () => {
-  const [popupVisible, setPopupVisible] = useState(false);
-  const [popupData, setPopupData] = useState(null);
-  const navigation = useNavigation();
-  
-  // if (!markers || markers.length === 0) 
-  //   {alert("Bombaclat");
-  //     return null;
-      
-  //   }
+  const { toggleModal, setModalData } = useContext(ModalContext); // Access setModalData
 
-  const handleMarkerPress = (marker) => {
-    setPopupData(marker);
-    setPopupVisible(true);
-  };
-
-  const closePopup = () => {
-    setPopupVisible(false);
-    setPopupData(null);
+  const handleMarkerPress = (building) => {
+    setModalData({
+      name: building.name,
+      coordinate: building.coordinate,
+      address: building.address,
+      fullBuildingName: building.fullBuildingName,
+    }); // Update modalData
+    toggleModal(); // Show modal
   };
 
   return (
     <>
-      {Building.map((marker) => (
+      {Building.map((building) => (
         <Marker
-          key={`${marker.name}-${marker.coordinate.latitude}-${marker.coordinate.longitude}`}
-          coordinate={marker.coordinate}
-          title={marker.name}
-          address={marker.address}
-          fullBuildingName={marker.fullBuildingName}
-          onPress={() => handleMarkerPress(marker)}
+          key={`${building.name}-${building.coordinate.latitude}-${building.coordinate.longitude}`}
+          testID={`marker-${building.name?.toLowerCase().replace(/\s+/g, "-") || building.id}`}
+          coordinate={building.coordinate}
+          title={building.name}
+          address={building.address}
+          fullBuildingName={building.fullBuildingName}
+          onPress={() => handleMarkerPress(building)}
         >
           <Image source={customMarkerImage} style={styles.markerImage} />
-          <Callout>
-            <View style={styles.calloutContainer}>
-              <Text style={styles.calloutText}>{marker.name}</Text>
-              <Text style={styles.calloutText}>{marker.address}</Text>
-              <Text style={styles.calloutText}>{marker.fullBuildingName}</Text>
-            </View>
-          </Callout>
         </Marker>
       ))}
-
-      {/* <PopupModal
-        isVisible={popupVisible}
-        data={popupData}
-        onClose={closePopup}
-        navigation = {navigation}
-      /> */}
     </>
   );
 };
 
-
-// MapMarkers.propTypes = {
-//   markers: PropTypes.arrayOf(
-//     PropTypes.shape({
-//       coordinate: PropTypes.shape({
-//         latitude: PropTypes.number.isRequired,
-//         longitude: PropTypes.number.isRequired,
-//       }).isRequired,
-//       name: PropTypes.string.isRequired,
-//     }),
-//   ),
-// };
-
 export default MapMarkers;
-
