+++ ConcordiaMaps/screen/HomeScreen.js
@@ -7,9 +7,6 @@ import TemporaryModal from "../components/temporaryModal";
 import { LocationContext } from "../contexts/LocationContext";
 import Footer from "../components/Footer";
 import styles from "../styles";
-const customMarkerImage = require("../assets/PinLogo.png");
-import { Building } from "../constants/Building";
-import { ModalContext } from "../App";
 import BuildingColoring from "../components/buildingColoring";
 import Legend from "../components/Legend";
 import ShuttleStop from "../components/ShuttleStop";
@@ -28,7 +25,6 @@ function HomeScreen({ asyncKey = "Campus" }) {
   const sgwPostalCode = process.env.EXPO_PUBLIC_SGW_POSTAL_CODE;
 
   const location = useContext(LocationContext);
-  const { toggleModal, setModalData } = useContext(ModalContext); // Access setModalData
 
   const [postalCode, setPostalCode] = useState(null);
   const [coordinates, setCoordinates] = useState(null);
@@ -114,17 +110,6 @@ function HomeScreen({ asyncKey = "Campus" }) {
     }, 100);
   };
 
-  // Function to handle marker press and pass data to the modal
-  const handleMarkerPress = (building) => {
-    setModalData({
-      name: building.name,
-      coordinate: building.coordinat,
-      address: building.address,
-      fullBuildingName: building.fullBuildingName,
-    }); // Update modalData
-    toggleModal(); // Show modal
-  };
-
   const [modalState, setModalState] = useState(true);
   useEffect(() => {
     if (modalState) {
@@ -176,23 +161,7 @@ function HomeScreen({ asyncKey = "Campus" }) {
             watchUserLocation={true}
             onRegionChangeComplete={(region) => setMapRegion(region)}
           >
-            {/* {Building.map((building) => (
-              <Marker
-                key={`${building.name}-${building.coordinate.latitude}-${building.coordinate.longitude}`}
-                testID={`marker-${building.name?.toLowerCase().replace(/\s+/g, "-") || building.id}`}
-                coordinate={building.coordinate}
-                title={building.name}
-                address={building.address}
-                fullBuildingName={building.fullBuildingName}
-                onPress={() => handleMarkerPress(building)}
-              >
-                <Image
-                  source={customMarkerImage}
-                  style={styles.customMarkerImage}
-                />
-              </Marker>
-            ))} */}
-            <MapMarkers/>
+            <MapMarkers />
             <BuildingColoring />
             {selectedLocation && (
               <Marker
